+++ CHANGELOG.md
@@ -2,6 +2,7 @@ v1.0.0-beta.3 [February 16, 2021](https://github.com/lando/vuepress-theme-defaul
 ---------------------------------
 
 * Improved `sidebarHeader.icon` so it honors `base`
+* Improved `socialLink` and `sidebarHeader` styling for mobile
 
 v1.0.0-beta.2 [February 16, 2021](https://github.com/lando/vuepress-theme-default-plus/releases/tag/v1.0.0-beta.2)
 ---------------------------------