+++ CHANGELOG.md
@@ -3,6 +3,7 @@ v1.0.0-beta.3 [February 16, 2021](https://github.com/lando/vuepress-theme-defaul
 
 * Improved `sidebarHeader.icon` so it honors `base`
 * Improved `social` and `sidebarHeader` styling for mobile
+* Improved `<form>` stylings
 * Fixed bug causing `lastUpdated` to not show up in `GuideHeader`
 
 v1.0.0-beta.2 [February 16, 2021](https://github.com/lando/vuepress-theme-default-plus/releases/tag/v1.0.0-beta.2)